♻️ Move encode_uid
