+++ bionty/_biorecord.py
@@ -2,72 +2,12 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
+from uids import encode_uid
+
 if TYPE_CHECKING:
     from types import ModuleType
 
     from .models import BioRecord
-
-
-def encode_uid(registry: type[BioRecord], kwargs: dict):
-    from lamindb.models import Record
-
-    from . import ids
-
-    if kwargs.get("uid") is not None:
-        # if uid is passed, no encoding is needed
-        return kwargs
-    name = registry.__name__.lower()
-    if hasattr(registry, "organism_id"):
-        organism = kwargs.get("organism")
-        if organism is None:
-            if kwargs.get("organism_id") is not None:
-                from .models import Organism
-
-                organism = Organism.get(kwargs.get("organism_id")).name
-        elif isinstance(organism, Record):
-            organism = organism.name
-    else:
-        organism = ""
-
-    if hasattr(registry, "_ontology_id_field"):
-        ontology_id_field = registry._ontology_id_field
-    else:
-        ontology_id_field = "ontology_id"
-    if hasattr(registry, "_name_field"):
-        name_field = registry._name_field
-    else:
-        name_field = "name"
-
-    str_to_encode = None
-    if name == "source":
-        str_to_encode = f"{kwargs.get('entity', '')}{kwargs.get('name', '')}{kwargs.get('organism', '')}{kwargs.get('version', '')}"
-    elif name == "gene":  # gene has multiple id fields
-        str_to_encode = kwargs.get(ontology_id_field)
-        if str_to_encode is None or str_to_encode == "":
-            str_to_encode = kwargs.get("stable_id")
-        if str_to_encode is None or str_to_encode == "":
-            str_to_encode = f"{kwargs.get(name_field)}{organism}"  # name + organism
-        if str_to_encode is None or str_to_encode == "":
-            raise AssertionError(
-                f"must provide {ontology_id_field}, stable_id or {name_field}"
-            )
-    else:
-        str_to_encode = kwargs.get(ontology_id_field)
-        if str_to_encode is None or str_to_encode == "":
-            str_to_encode = f"{kwargs.get(name_field)}{organism}"  # name + organism
-        if str_to_encode is None or str_to_encode == "":
-            raise AssertionError(f"must provide {ontology_id_field} or {name_field}")
-
-    if str_to_encode is not None and len(str_to_encode) > 0:
-        try:
-            id_encoder = getattr(ids, name)
-        except Exception:
-            if ontology_id_field == "ontology_id":
-                id_encoder = ids.ontology
-            else:
-                return kwargs
-        kwargs["uid"] = id_encoder(str_to_encode)
-    return kwargs
 
 
 def lookup2kwargs(record: BioRecord, *args, **kwargs) -> dict:
